# packages/openapi/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Poveroh API",
        "version": "1.0.0",
        "description": "OpenAPI definition for the Poveroh v1 API"
    },
    "servers": [
        {
            "url": "http://localhost:3001/v1",
            "description": "API Poveroh Localhost v1"
        },
        {
            "url": "http://api.poveroh.local/v1",
            "description": "API Poveroh Localhost proxy v1"
        },
        {
            "url": "http://api.poveroh.com/v1",
            "description": "API Poveroh Production v1"
        }
    ],
    "components": {
        "schemas": {
...
            "StringFilter": {
                "type": "object",
                "properties": {
                    "equals": {
                        "type": "string"
                    },
                    "contains": {
                        "type": "string"
                    }
                }
            },
...
            "FilterOptions": {
                "type": "object",
                "properties": {
                    "skip": {
                        "type": "integer",
                        "minimum": 0,
                        "default": 0
                    },
                    "take": {
                        "type": "integer",
                        "minimum": 0,
                        "exclusiveMinimum": true,
                        "default": 20
                    },
                    "sortBy": {
                        "type": "string",
                        "default": "createdAt"
                    },
                    "sortOrder": {
                        "type": "string",
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "default": "desc"
                    }
                }
            },
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            false
                        ],
                        "description": "Always false for error responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Error message describing what went wrong"
                    },
                    "error": {
                        "nullable": true,
                        "description": "Optional additional error details or context"
                    }
                },
                "required": [
                    "success",
                    "message"
                ]
            },
...
            "GetSubcategoryListResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "description": "Always true for success responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Optional success message"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SubcategoryDataResponse"
                        },
                        "description": "Response data"
                    }
                },
                "required": [
                    "success"
                ]
            },
            "GetSubcategoryResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "description": "Always true for success responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Optional success message"
                    },
                    "data": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/SubcategoryDataResponse"
                            },
                            {
                                "description": "Response data"
                            }
                        ]
                    }
                },
                "required": [
                    "success"
                ]
            },
            "CreateSubcategoryRequest": {
                "type": "object",
                "properties": {
                    "categoryId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "logoIcon": {
                        "type": "string",
                        "format": "uri"
                    }
                },
                "required": [
                    "categoryId",
                    "title",
                    "logoIcon"
                ]
            },
            "CreateSubcategoryMultipartRequest": {
                "type": "object",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/CreateSubcategoryRequest"
                    },
                    "file": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "binary",
                            "description": "Optional file upload"
                        }
                    }
                },
                "required": [
                    "file"
                ]
            },
            "CreateSubcategoryResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "description": "Always true for success responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Optional success message"
                    },
                    "data": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/SubcategoryDataResponse"
                            },
                            {
                                "description": "Response data"
                            }
                        ]
                    }
                },
                "required": [
                    "success"
                ]
            },
            "UpdateSubcategoryRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "logoIcon": {
                        "type": "string",
                        "format": "uri"
                    }
                }
            },
            "UpdateSubcategoryResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "description": "Always true for success responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Optional success message"
                    },
                    "data": {
                        "nullable": true,
                        "description": "Response data"
                    }
                },
                "required": [
                    "success"
                ]
            },
            "DeleteSubcategoryResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "description": "Always true for success responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Optional success message"
                    },
                    "data": {
                        "nullable": true,
                        "description": "Response data"
                    }
                },
                "required": [
                    "success"
                ]
            },
            "SubcategoryParamsId": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "id"
                ]
            },
            "SubcategoryFilters": {
                "type": "object",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/SubcategoryParamsId"
                    },
                    "title": {
                        "$ref": "#/components/schemas/StringFilter"
                    },
                    "description": {
                        "$ref": "#/components/schemas/StringFilter"
                    },
                    "categoryId": {
                        "type": "object",
                        "properties": {
                            "categoryId": {
                                "type": "string",
                                "format": "uuid"
                            }
                        },
                        "required": [
                            "categoryId"
                        ]
                    }
                },
                "additionalProperties": {
                    "anyOf": [
                        {
                            "type": "string"
                        },
                        {
                            "$ref": "#/components/schemas/StringFilter"
                        }
                    ]
                }
            },
...
            "SubcategoryDataResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "categoryId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "logoIcon": {
                        "type": "string",
                        "format": "uri"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "categoryId",
                    "title",
                    "logoIcon",
                    "createdAt",
                    "updatedAt"
                ]
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/subcategories": {
            "get": {
                "tags": [
                    "Subcategory"
                ],
                "operationId": "getSubcategories",
                "summary": "Get all subcategories",
                "description": "Retrieve a list of all subcategories associated with the user",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/components/schemas/SubcategoryFilters"
                        },
                        "required": false,
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "schema": {
                            "$ref": "#/components/schemas/FilterOptions"
                        },
                        "required": false,
                        "name": "options",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of subcategories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetSubcategoryListResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subcategory not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Subcategory"
                ],
                "operationId": "createSubcategory",
                "summary": "Create subcategory",
                "description": "Create a new subcategory with the provided data",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "description": "Subcategory data to create",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSubcategoryMultipartRequest"
                            },
                            "encoding": {
                                "data": {
                                    "contentType": "application/json"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subcategory created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateSubcategoryResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Subcategory"
                ],
                "operationId": "deleteSubcategories",
                "summary": "Delete all subcategories",
                "description": "Delete all subcategories associated with the user",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subcategories deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteSubcategoryResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/subcategories/{id}": {
            "get": {
                "tags": [
                    "Subcategory"
                ],
                "operationId": "getSubcategoryById",
                "summary": "Get subcategory by ID",
                "description": "Retrieve a specific subcategory by its ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true,
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subcategory found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetSubcategoryResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subcategory not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Subcategory"
                ],
                "operationId": "updateSubcategory",
                "summary": "Update subcategory",
                "description": "Update an existing subcategory with the provided data",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true,
                        "name": "id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "description": "Subcategory data to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSubcategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subcategory updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateSubcategoryResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subcategory not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Subcategory"
                ],
                "operationId": "deleteSubcategory",
                "summary": "Delete subcategory",
                "description": "Delete an existing subcategory with the provided ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true,
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subcategory deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteSubcategoryResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subcategory not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}